Revert "Рефактор моделей и миграций"
