+++ app/Models/User.php
@@ -12,13 +12,17 @@ class User extends Authenticatable
 {
     use HasApiTokens, HasFactory, Notifiable;
 
-    protected $table = 'users';
     /**
      * The attributes that are mass assignable.
      *
      * @var array<int, string>
      */
-    protected $fillable = ['name', 'surname', 'email', 'birthdate', 'position'];
+    protected $fillable = [
+        'name',
+        'email',
+        'password',
+    ];
+
     /**
      * The attributes that should be hidden for serialization.
      *
@@ -38,9 +42,4 @@ class User extends Authenticatable
         'email_verified_at' => 'datetime',
         'password' => 'hashed',
     ];
-
-    public function tasks()
-    {
-        return $this->hasMany(Task::class, 'user_id', 'id');
-    }
 }
