+++ app/Models/Worker.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Worker extends Model
+{
+    use HasFactory;
+
+    protected $table = 'workers';
+
+    protected $fillable = ['name', 'surname', 'email', 'birthdate', 'position'];
+
+    public function tasks()
+    {
+        return $this->hasMany(Task::class, 'worker_id', 'id');
+    }
+}
