+++ database/factories/UserFactory.php
@@ -22,13 +22,10 @@ class UserFactory extends Factory
     {
         return [
             'name' => fake()->name(),
-            'surname' => fake()->lastName,
             'email' => fake()->unique()->safeEmail(),
-            'birthdate' => fake()->date,
             'email_verified_at' => now(),
             'password' => static::$password ??= Hash::make('password'),
             'remember_token' => Str::random(10),
-            'position' => fake()->jobTitle,
         ];
     }
