+++ database/factories/WorkerFactory.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Database\Factories;
+
+use App\Models\Task;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Worker>
+ */
+class WorkerFactory extends Factory
+{
+    /**
+     * Define the model's default state.
+     *
+     * @return array<string, mixed>
+     */
+    public function definition(): array
+    {
+        return [
+            'name' => fake()->name,
+            'surname' => fake()->lastName,
+            'email' => fake()->email,
+            'birthdate' => fake()->date,
+            'position' => fake()->jobTitle,
+        ];
+    }
+}
